requirements.txt
tesseract
pdf2image